+++ nflow-engine/src/main/java/io/nflow/engine/internal/executor/WorkflowStateProcessor.java
@@ -22,6 +22,7 @@ import java.util.EnumSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.Random;
 
 import org.joda.time.DateTime;
 import org.joda.time.Duration;
@@ -74,6 +75,7 @@ class WorkflowStateProcessor implements Runnable {
   private final Map<Integer, WorkflowStateProcessor> processingInstances;
   private long startTimeSeconds;
   private Thread thread;
+  private Random rnd = new Random();
 
   WorkflowStateProcessor(int instanceId, ObjectStringMapper objectMapper, WorkflowDefinitionService workflowDefinitions,
       WorkflowInstanceService workflowInstances, WorkflowInstanceDao workflowInstanceDao,
@@ -151,6 +153,7 @@ class WorkflowStateProcessor implements Runnable {
           processAfterFailureListeners(listenerContext, execution.getThrown());
         } else {
           processAfterListeners(listenerContext);
+          optionallyCleanupWorkflowInstanceHistory(definition.getSettings());
         }
         subsequentStateExecutions = busyLoopPrevention(state, settings, subsequentStateExecutions, execution);
         instance = saveWorkflowInstanceState(execution, instance, definition, actionBuilder);
@@ -297,6 +300,17 @@ class WorkflowStateProcessor implements Runnable {
       return nextAction;
     }
     return new SkippedStateHandler(nextAction, instance, definition, execution, state).processState();
+  }
+
+  private void optionallyCleanupWorkflowInstanceHistory(WorkflowSettings settings) {
+    if (settings.historyDeletableAfterHours != null && roughlyEveryTenthTime()) {
+      logger.debug("Cleaning workflow history older than {} hours", settings.historyDeletableAfterHours);
+      workflowInstanceDao.deleteWorkflowInstanceHistory(instanceId, settings.historyDeletableAfterHours);
+    }
+  }
+
+  private boolean roughlyEveryTenthTime() {
+    return rnd.nextInt(10) == 0;
   }
 
   static class ExecutorListenerChain implements ListenerChain {
@@ -542,5 +556,9 @@ class WorkflowStateProcessor implements Runnable {
       sb.append(element).append('\n');
     }
     return sb;
+  }
+
+  void setRandom(Random rnd) {
+    this.rnd = rnd;
   }
 }