+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 **Details**
 - Popup notifications from workflow instance updates, network and authentication issues in Explorer
 - Upgraded Spring to version 5.1.3.RELEASE
+- Workflow instance history (actions, states) cleanup as part of state processing, configurable through `WorkflowSettings`
 
 ## 5.2.0 (2018-11-20)